Flyttet VilkårGruppeConfig till egen fil og flyttet ut metode fra klassen i Unntak
